[api] feat: nothing() now compensates for Maybe[Foo]
nothing(::Type) should be called with the type
you want wrapped with Maybe. For example

nothing(Int) not nothing(Maybe[Int])

However, I kept calling it with the latter, and
it seems silly to punish people for such a
reasonable "mistake". It's easy enough to tell
what they meant.

TAGS:  enhancement